refactor: services/ml_models
